+++ stories/DeprecatedAdminPanel.stories.ts
@@ -27,6 +27,6 @@ type Story = StoryObj<typeof meta>;
 
 export const Loading: Story = {
   args: {
-    room: 'storybook',
+    room: 'default',
   },
 };